:hammer: Move prop hydration functions.
