+++ src/renderer/js/components/Updater.jsx
@@ -1,127 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {merge, type, omit, map} from 'ramda';
-import Wrapper from './Wrapper';
+import {map} from 'ramda';
 import {apiRequest} from '../requests';
 import {loadCss, loadScript} from '../../../commons/js';
-
-function isComponent(c) {
-    return (
-        type(c) === 'Object' &&
-        (c.hasOwnProperty('package') &&
-            c.hasOwnProperty('aspects') &&
-            c.hasOwnProperty('name') &&
-            c.hasOwnProperty('identity'))
-    );
-}
-
-function hydrateProps(props, updateAspects, connect, disconnect) {
-    const replace = {};
-    Object.entries(props).forEach(([k, v]) => {
-        if (type(v) === 'Array') {
-            replace[k] = v.map(c => {
-                if (!isComponent(c)) {
-                    // Mixing components and primitives
-                    return c;
-                }
-                const newProps = hydrateProps(
-                    c.aspects,
-                    updateAspects,
-                    connect,
-                    disconnect
-                );
-                if (!newProps.key) {
-                    newProps.key = c.identity;
-                }
-                return hydrateComponent(
-                    c.name,
-                    c.package,
-                    c.identity,
-                    newProps,
-                    updateAspects,
-                    connect,
-                    disconnect
-                );
-            });
-        } else if (isComponent(v)) {
-            const newProps = hydrateProps(
-                v.aspects,
-                updateAspects,
-                connect,
-                disconnect
-            );
-            replace[k] = hydrateComponent(
-                v.name,
-                v.package,
-                v.identity,
-                newProps,
-                updateAspects,
-                connect,
-                disconnect
-            );
-        } else if (type(v) === 'Object') {
-            replace[k] = hydrateProps(v, updateAspects, connect, disconnect);
-        }
-    });
-    return merge(props, replace);
-}
-
-function hydrateComponent(
-    name,
-    package_name,
-    identity,
-    props,
-    updateAspects,
-    connect,
-    disconnect
-) {
-    const pack = window[package_name];
-    const element = React.createElement(pack[name], props);
-    return (
-        <Wrapper
-            identity={identity}
-            updateAspects={updateAspects}
-            component={element}
-            connect={connect}
-            package_name={package_name}
-            component_name={name}
-            aspects={props}
-            disconnect={disconnect}
-            key={`wrapper-${identity}`}
-        />
-    );
-}
-
-function prepareProp(prop) {
-    if (React.isValidElement(prop)) {
-        return {
-            identity: prop.props.identity,
-            aspects: map(
-                prepareProp,
-                omit(
-                    [
-                        'identity',
-                        'updateAspects',
-                        '_name',
-                        '_package',
-                        'aspects',
-                        'key',
-                    ],
-                    prop.props.aspects // You actually in the wrapper here.
-                )
-            ),
-            name: prop.props.component_name,
-            package: prop.props.package_name,
-        };
-    }
-    if (type(prop) === 'Array') {
-        return prop.map(prepareProp);
-    }
-    if (type(prop) === 'Object') {
-        return map(prepareProp, prop);
-    }
-    return prop;
-}
+import {
+    hydrateComponent,
+    hydrateProps,
+    isComponent,
+    prepareProp,
+} from '../hydrator';
 
 export default class Updater extends React.Component {
     constructor(props) {
